Proteção no vanish.sh

+++ scripts/vanish.sh
@@ -1,3 +1,9 @@
+# Proteção
+if [ "$1" != "--confirm" ]; then
+    echo "Error: Please pass --confirm to run this script."
+    exit 1
+fi
+
 # Remove containers e volumes
 # shellcheck disable=SC2046
 docker rm -vf $(docker ps -a -q) 2> /dev/null
